[ASTImporter] Separate test files
